refactor: rename title.ts -> titleContext.ts
